+++ src/main/java/com/tamstar/pensionchats/core/service/ProprietaireService.java
@@ -17,7 +17,7 @@ public class ProprietaireService {
 		proprietaireRepository.create(proprietaire);
 	}
 
-	public Proprietaire getProprietaire(Long id) {
+	public Proprietaire getProprietaire(Short id) {
 		return proprietaireRepository.getById(id);
 	}
